Convert `create-component` to TypeScript and simplify types
